requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium==4.15.2
pandas>=2.0.0
huggingface-hub>=0.19.4
tqdm>=4.66.1
python-dotenv>=1.0.0
webdriver-manager==4.0.1
numpy==1.24.3
scipy==1.10.1

# Document processing
PyPDF2==3.0.1
python-docx>=0.8.11
markdown>=3.5.1
pdfplumber>=0.10.2
tabula-py>=2.8.0
openpyxl>=3.1.2

# Scientific formats
astropy>=5.3.0
h5py>=3.10.0
netCDF4>=1.6.5
cfgrib>=0.9.10
vtk>=9.3.0
PyMuPDF>=1.23.7  # fitz

# Data formats
msgpack>=1.0.7
PyYAML>=6.0.1
toml>=0.10.2
epub>=0.5.2
python-pptx>=0.6.21

# Image processing
Pillow>=10.1.0
pytesseract>=0.3.10

# Scientific APIs
arxiv>=1.4.8
ads>=0.12.3
astroquery>=0.4.6
crossref-commons>=0.0.7
doi2bib>=0.4.0
scholarly>=1.7.11
feedparser>=6.0.10

# Utilities
chardet>=5.2.0
latex2mathml>=3.76.0
nbformat>=5.9.2
nbconvert>=7.11.0
retry>=0.9.2
aiohttp>=3.9.1
asyncio>=3.4.3
tenacity>=8.2.3

# Core Machine Learning Dependencies
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
scibert==0.5.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.3
gensim==4.3.2

# Deep Learning and Embedding
sentence-transformers==2.3.1
allennlp==2.10.0
allennlp-models==2.10.0

# Knowledge Graph and Reasoning
networkx==3.2.1
rdflib==7.0.0

# Knowledge Integration Dependencies
networkx==3.2.1
pandas==2.1.4
transformers==4.36.2
deepset-transformers==0.4.0

# Performance and Optimization
numba==0.58.1
cupy==12.2.0  # GPU acceleration
faiss-cpu==1.7.4

# Performance Metrics and Evaluation
scikit-learn==1.3.2
numpy==1.24.3

# Logging and Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing and Validation
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Configuration Management
pyyaml==6.0.1
hydra-core==1.3.2

# Development and Debugging
ipython==8.18.1
jupyter==1.0.0

# Optional: Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Optional: Advanced NLP
spacy-transformers==1.2.5
stanza==1.5.0

# Optional: Quantum-Inspired Computing
pennylane==0.33.0
qiskit==0.43.0

# Experimental: AI Reasoning
sympy==1.12
z3-solver==4.12.2

# Experimental: Explainable AI
shap==0.43.0
lime==0.2.0.1

# Optional: GPU Support
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118  # Optional CUDA version

# Meta-Learning and Few-Shot Learning
higher==0.2.0
learn2learn==0.2.1
higher-order-layers==0.1.3

# Advanced Fine-Tuning and Optimization
torch-optimizer==0.3.0
pytorch-lightning==2.1.2

# Natural Language Interface Dependencies
spacy==3.7.2
langdetect==1.0.9
dash==2.14.2
plotly==5.18.0
dash-core-components==2.0.0
dash-html-components==2.0.0

# Multilingual NLP Models
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl

# Advanced Semantic Understanding
transformers==4.36.2
torch-nlp==0.2.0
sentence-transformers==2.3.1
bart-large-mnli==0.1.0

# Personalized Interaction
scikit-learn==1.3.2
scipy==1.11.4

# Distributed Knowledge Graph Dependencies
ray[default]==2.9.0
dask==2024.1.0
distributed==2024.1.0
networkx==3.2.1

# Quantum-Inspired Computation
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Advanced Reasoning Dependencies
scipy==1.11.4
networkx==3.2.1

# Probabilistic Reasoning
pomegranate==0.14.7
pgmpy==0.1.23

# Hypothesis Generation
numpy==1.26.2
scikit-learn==1.3.2

# Meta-Cognitive Reasoning Dependencies
numpy==1.26.2
scipy==1.11.4

# Cognitive Modeling
pymc3==3.11.4
arviz==0.17.1

# Advanced Probabilistic Reasoning
probables==0.5.6
pyro-ppl==1.8.4

# Visualization Dependencies
matplotlib==3.8.2
networkx==3.2.1

# Explainable AI
shap==0.43.0
lime==0.2.0.1

# Advanced Visualization
plotly==5.18.0
seaborn==0.13.1

# Semantic Understanding
transformers==4.36.2
sentence-transformers==2.3.1

# Advanced Language Models
flair==0.12.2
stanza==1.6.0

# Multilingual Support
polyglot==16.7.4
langdetect==1.0.9

# Research Generation Dependencies
scikit-learn==1.3.2
scipy==1.11.4

# Research Analytics
bibliopixel==3.1.0
scholarly==0.7.2

# Bibliometric Analysis
pybtex==0.24.0
habanero==0.8.0

# Research Trend Analysis
altmetric==1.0.3
crossref-commons==0.4.7
